split Containers presentation implementation to appropriate files
