+++ applications/launchpad/gui-react/src/containers/Dashboard/ExpertView/Containers/styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const ContainersTable = styled.table`
+  width: 100%;
+  max-width: 50vw;
+  margin-top: ${({ theme }) => theme.spacing()};
+`
+
+export const TdRight = styled.td`
+  text-align: right;
+`
