+++ applications/launchpad/gui-react/src/containers/Dashboard/ExpertView/Containers/types.ts
@@ -1,0 +1,1 @@
+export type ServiceDto = any
